Refactor driver ABI and provider descriptor structures
- Updated the `DriverVTable` initialization to use a struct for function pointers, improving readability and maintainability.
- Changed the `abi_size` field in `ProviderDescriptorV1` from `u32` to `u64` to accommodate larger sizes.
- Adjusted various instances of `abi_size` checks to use `u64` for consistency.
- Modified the `recv_payload` method in `TcpStream` to use a more concise pattern matching.
- Cleaned up unnecessary `clippy` allows in `lib.rs` and `security.rs`.
- Enhanced error handling in `CapabilityManager` methods with detailed documentation.
- Updated QEMU runner to pass configuration by reference, improving performance and clarity.

+++ drivers/example_abi/src/lib.rs
@@ -68,7 +68,7 @@ extern "C" fn driver_fini() -> i32 {
 #[unsafe(no_mangle)]
 pub static DRIVER_EXPORTS: DriverExportsV1 = DriverExportsV1 {
     abi_version: DRIVER_EXPORTS_ABI_VERSION,
-    abi_size: core::mem::size_of::<DriverExportsV1>() as u32,
+    abi_size: core::mem::size_of::<DriverExportsV1>() as u64,
     name_ptr: driver_name().as_ptr(),
     name_len: driver_name().len(),
     entry: _exorust_driver_entry,
